refactor: move chunk marshalling to cmd
